refact: recorder-service 책임 분리
